🐛 fix: workspace store subscribe로 변경 감지 후 저장 호출
